fix!: Remove <Leader> autoprefix on key_mapping
This removes the unconditional `<Leader>` prefix when `key_mapping`
(default `"jq"`) is added into the keymap. From now on, user config
must include `<Leader>` explicitly if `key_mapping` is overridden.

This also removes the block that sets `<Leader>` to `<Space>` if unset,
since this no longer depends on it at all. That change also avoids some
unexpected behavior if `vim.g.mapleader` is nil (which defaults to
backslash, not space!).